Socials Jobs, Config
