+++ app/jobs/socials/telegram_post_job.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+# Job for send message to telegram channel (gem 'telegram-bot-ruby')
+module Socials
+  # TODO: documentation is missing for this class
+  # We should consider addig some documentation here
+  class TelegramPostJob < BasePostJob
+    require 'telegram/bot'
+
+    def perform(job_id)
+      super
+
+      prepare_for('TELEGRAM')
+
+      Telegram::Bot::Client.run(token) do |bot|
+        bot.api.send_message(chat_id: group_id, text: message)
+      end
+    end
+  end
+end
