+++ app/services/social_web_poster_service.rb
@@ -3,9 +3,9 @@
 # Service for start all jobs for pubslish
 class SocialWebPosterService
   def self.post_job(job_id)
-    VkontaktePostJob.perform_later(job_id)
-    FacebookPostJob.perform_later(job_id)
-    TelegramPostJob.perform_later(job_id)
-    TwitterPostJob.perform_later(job_id)
+    Socials::VkontaktePostJob.perform_later(job_id)
+    Socials::FacebookPostJob.perform_later(job_id)
+    Socials::TelegramPostJob.perform_later(job_id)
+    Socials::TwitterPostJob.perform_later(job_id)
   end
 end
