+++ spec/jobs/social_web_poster_service_spec.rb
@@ -7,15 +7,18 @@ RSpec.describe SocialWebPosterService do
   let(:post) { SocialWebPosterService.post_job(1) }
 
   it 'VkontakteWorker is working' do
-    expect { post }.to have_enqueued_job(VkontaktePostJob)
+    expect { post }.to have_enqueued_job(Socials::VkontaktePostJob)
   end
+
   it 'FacebookWorker is working' do
-    expect { post }.to have_enqueued_job(FacebookPostJob)
+    expect { post }.to have_enqueued_job(Socials::FacebookPostJob)
   end
+
   it 'TelegramWorker is working' do
-    expect { post }.to have_enqueued_job(TelegramPostJob)
+    expect { post }.to have_enqueued_job(Socials::TelegramPostJob)
   end
+
   it 'TwitterWorker is working' do
-    expect { post }.to have_enqueued_job(TwitterPostJob)
+    expect { post }.to have_enqueued_job(Socials::TwitterPostJob)
   end
 end
